refactor(project): updated clean.sh + minor adjustments of configs

+++ clean.sh
@@ -1,14 +1,18 @@
 #!/bin/bash
 
 if [[ $EUID -ne 0 ]]; then
-    echo "This script requires root privileges"
-    exit 1
+  echo "This script requires root privileges"
+  exit 1
+fi
+
+if grep -q "boot=overlay" /proc/cmdline; then
+  echo "OverlayFS is still enabled!"
+  exit 1
 fi
 
 # remove development files
 rm -f /home/modbros/mobro-raspberrypi/dependencies.txt
 rm -f /home/modbros/mobro-raspberrypi/mobro-raspberrypi.iml
-rm -f /home/modbros/mobro-raspberrypi/install.sh
 rm -rf /home/modbros/mobro-raspberrypi/.idea
 
 # remove documentation from already installed packages
@@ -20,7 +24,6 @@ rm -f /var/log/{auth,boot,bootstrap,daemon,kern}.log
 rm -f /var/log/{debug,dmesg,messages,syslog}
 
 rm -f /home/modbros/mobro-raspberrypi/log/log_?.txt
-: >/home/modbros/mobro-raspberrypi/log/log.txt
 
 # clean apt cache
 apt-get clean
@@ -28,11 +31,8 @@ apt-get clean
 # reset wpa config
 cp -f /home/modbros/mobro-raspberrypi/config/wpa_supplicant_clean.conf /etc/wpa_supplicant/wpa_supplicant.conf
 
-# reset data files
-: >/home/modbros/mobro-raspberrypi/data/hosts
-: >/home/modbros/mobro-raspberrypi/data/ssids
-cat /home/modbros/mobro-raspberrypi/config/mobro_config_default.txt >/home/modbros/mobro-raspberrypi/data/mobro_config
-echo "0" >/home/modbros/mobro-raspberrypi/data/mobro_found
+# reset config file
+cat /home/modbros/mobro-raspberrypi/config/mobro_config_default >/home/modbros/mobro-raspberrypi/config/mobro_config
 
 # delete cache + bash history
 rm -rf /home/modbros/.cache/*
@@ -43,5 +43,7 @@ cat /home/modbros/mobro-raspberrypi/config/config.txt >/boot/config.txt
 cat /home/modbros/mobro-raspberrypi/config/99-fbturbo.conf >/usr/share/X11/xorg.conf.d/99-fbturbo.conf
 sudo rm -rf /etc/X11/xorg.conf.d/*
 
-# overwrite free space of partition
+# overwrite free space of partitions
 #sfill -f -z -l -l -v /
+#sfill -f -z -l -l -v /boot
+#sfill -f -z -l -l -v /home
